mcp/client/requirements.txt
fastapi
uvicorn
mcp
openai
python-dotenv
